refactor: split auth data source into local and remote
