+++ lib/features/user/data/datasources/remote/auth_remote_data_source_impl.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import '../../../../core/core.dart';
-import '../../domain/entities/profile_entity.dart';
-import '../models/profile_model.dart';
+import '../../../../../core/core.dart';
+import '../../../domain/entities/profile_entity.dart';
+import '../../models/profile_model.dart';
 import 'auth_remote_data_source.dart';
 
 class AuthRemoteDataSourceImpl implements AuthRemoteDataSource {
